requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==4.2.7
django-appconf==1.0.5
django-crispy-forms==2.1
django-extensions==3.2.3
django-imagekit==5.0.0
pilkit==3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3